pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-autodrive_env"
version = "0.1.0"
description = "AutoDrive Gym: OpenEnv driving environment for Indian road reasoning."
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.0",
    "openai>=1.0.0",
    "requests>=2.25.0",
    "fastapi",
    "uvicorn",
    "pydantic",
]

[project.optional-dependencies]
train = [
    "torch",
    "peft",
    "transformers",
    "datasets",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "autodrive_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["autodrive_env", "autodrive_env.server"]
package-dir = { "autodrive_env" = ".", "autodrive_env.server" = "server" }
